fix examples for renamed methods toArray --> toDArray/toVArray
