--- api-examples/class.HH.ImmMap/toArray/001-basic-usage.md
@@ -1,1 +1,0 @@
-See [`Map::toArray`](/hack/reference/class/Map/toArray/#examples) for usage examples.
