+++ api-examples/class.HH.ImmMap/toDArray/001-basic-usage.md
@@ -1,0 +1,1 @@
+See [`Map::toDArray`](/hack/reference/class/Map/toDArray/#examples) for usage examples.
