--- api-examples/class.HH.ImmSet/toArray/001-basic-usage.md
@@ -1,1 +1,0 @@
-See [`Set::toArray`](/hack/reference/class/Set/toArray/#examples) for usage examples.
