+++ api-examples/class.HH.ImmSet/toDArray/001-basic-usage.md
@@ -1,0 +1,1 @@
+See [`Set::toDArray`](/hack/reference/class/Set/toDArray/#examples) for usage examples.
