+++ api-examples/class.HH.ImmSet/toVArray/001-basic-usage.md
@@ -1,0 +1,1 @@
+See [`Set::toVArray`](/hack/reference/class/Set/toVArray/#examples) for usage examples.
