--- api-examples/class.HH.ImmVector/toArray/001-basic-usage.md
@@ -1,1 +1,0 @@
-See [`Vector::toArray`](/hack/reference/class/Vector/toArray/#examples) for usage examples.
