+++ api-examples/class.HH.ImmVector/toVArray/001-basic-usage.md
@@ -1,0 +1,1 @@
+See [`Vector::toVArray`](/hack/reference/class/Vector/toVArray/#examples) for usage examples.
