+++ api-examples/class.HH.Map/toDArray/001-basic-usage.php.hhvm.expect
