+++ api-examples/class.HH.Pair/toVArray/001-basic-usage.php.hhvm.expect
