+++ api-examples/class.HH.Pair/toVArray/001-basic-usage.php.typechecker.expect
