+++ api-examples/class.HH.Set/toDArray/001-basic-usage.php
@@ -1,14 +1,13 @@
 <?hh // partial
 
-namespace Hack\UserDocumentation\API\Examples\Set\ToArray;
+namespace Hack\UserDocumentation\API\Examples\Set\ToDArray;
 
+<<__EntryPoint>>
 function run() {
   $s = Set {'red', 'green', 'blue', 'yellow'};
 
-  $array = $s->toArray();
+  $array = $s->toDArray();
 
   \var_dump(\is_array($array));
   \var_dump($array);
 }
-
-run();
