+++ api-examples/class.HH.Set/toDArray/001-basic-usage.php.hhvm.expect
