+++ api-examples/class.HH.Set/toVArray/001-basic-usage.php.typechecker.expect
