+++ api-examples/class.HH.Shapes/toArray/001-basic-usage.php
@@ -2,9 +2,8 @@
 
 namespace Hack\UserDocumentation\API\Examples\HH\Shapes\ToArray;
 
+<<__EntryPoint>>
 function run(): void {
   $point = shape('name' => 'Jane Doe', 'age' => 55, 'points' => 25.30);
   \var_dump(Shapes::toArray($point));
 }
-
-run();
