--- api-examples/class.HH.Vector/toArray/001-basic-usage.php
@@ -1,10 +1,0 @@
-<?hh // partial
-
-namespace Hack\UserDocumentation\API\Examples\Vector\ToArray;
-
-$v = Vector {'red', 'green', 'blue', 'yellow'};
-
-$array = $v->toArray();
-
-var_dump(is_array($array));
-var_dump($array);
