+++ api-examples/class.HH.Vector/toVArray/001-basic-usage.php.hhvm.expect
@@ -1,0 +1,11 @@
+bool(true)
+array(4) {
+  [0]=>
+  string(3) "red"
+  [1]=>
+  string(5) "green"
+  [2]=>
+  string(4) "blue"
+  [3]=>
+  string(6) "yellow"
+}
