+++ api-examples/class.HH.Vector/toVArray/001-basic-usage.php.typechecker.expect
@@ -1,0 +1,1 @@
+No errors!
